Endrer JaNeiSpørsmål til å ta inn ISvar..
... i stedet for boolean
- Flytter på utils til mer tilhørende filer
