+++ src/helpers/omdeg.ts
@@ -1,0 +1,24 @@
+export const hentSivilstatus = (statuskode: string) => {
+  switch (statuskode) {
+    case 'GIFT':
+      return 'Gift';
+
+    case 'UGIF':
+      return 'Ugift';
+
+    case 'SAMB':
+      return 'Samboer';
+
+    case 'SEPA':
+      return 'Separert';
+
+    case 'SKIL':
+      return 'Skilt';
+
+    case 'ENKE':
+      return 'Enke/ enkemann';
+
+    default:
+      return 'Annen sivilstatus enn GIFT, UGIF, SAMB, SEPA, SKIL';
+  }
+};
