+++ src/søknad/steg/1-omdeg/personopplysninger/Personopplysninger.tsx
@@ -11,8 +11,9 @@ import SeksjonGruppe from '../../../../components/gruppe/SeksjonGruppe';
 import JaNeiSpørsmål from '../../../../components/spørsmål/JaNeiSpørsmål';
 import { borDuPåDenneAdressen } from './PersonopplysningerConfig';
 import AlertStripe from 'nav-frontend-alertstriper';
-import { ISpørsmål } from '../../../../models/spørsmalogsvar';
+import { ISpørsmål, ISvar } from '../../../../models/spørsmalogsvar';
 import Lenke from 'nav-frontend-lenker';
+import { hentBooleanFraValgtSvar } from '../../../../utils/spørsmålogsvar';
 
 const Personopplysninger: React.FC = () => {
   const intl = useIntl();
@@ -25,7 +26,12 @@ const Personopplysninger: React.FC = () => {
     undefined
   );
 
-  const settPersonopplysningerFelt = (spørsmål: ISpørsmål, svar: boolean) => {
+  const settPersonopplysningerFelt = (
+    spørsmål: ISpørsmål,
+    valgtSvar: ISvar
+  ) => {
+    const svar: boolean = hentBooleanFraValgtSvar(valgtSvar);
+
     settSøknad({
       ...søknad,
       søkerBorPåRegistrertAdresse: { label: spørsmål.søknadid, verdi: svar },
