+++ src/søknad/steg/4-barnasbosted/BarnetsBostedEndre.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import useSøknadContext from '../../../context/SøknadContext';
-import { ISpørsmål } from '../../../models/spørsmalogsvar';
+import { ISpørsmål, ISvar } from '../../../models/spørsmalogsvar';
 import { boddSammenFør, borISammeHus, hvorMyeSammen } from './ForeldreConfig';
 import { IForelder, IBarn } from '../../../models/person';
 import JaNeiSpørsmål from '../../../components/spørsmål/JaNeiSpørsmål';
@@ -16,6 +16,7 @@ import OmAndreForelder from './OmAndreForelder';
 import BostedOgSamvær from './BostedOgSamvær';
 import SkalBarnBoHosDeg from './SkalBarnBoHosDeg';
 import AnnenForelderKnapper from './AnnenForelderKnapper';
+import { hentBooleanFraValgtSvar } from '../../../utils/spørsmålogsvar';
 
 interface Props {
   barn: IBarn;
@@ -44,10 +45,12 @@ const BarnetsBostedEndre: React.FC<Props> = ({
     // eslint-disable-next-line
   }, []);
 
-  const settHarBoddsammenFør = (spørsmål: ISpørsmål, valgtSvar: boolean) => {
-    const nyForelder = { ...forelder, [boddSammenFør.søknadid]: valgtSvar };
+  const settHarBoddsammenFør = (spørsmål: ISpørsmål, valgtSvar: ISvar) => {
+    const svar: boolean = hentBooleanFraValgtSvar(valgtSvar);
 
-    if (valgtSvar === false) {
+    const nyForelder = { ...forelder, [boddSammenFør.søknadid]: svar };
+
+    if (svar === false) {
       delete nyForelder.flyttetFra;
     }
