+++ src/søknad/steg/5-aktivitet/underUtdanning/TidligereUtdanning.tsx
@@ -12,10 +12,11 @@ import SeksjonGruppe from '../../../../components/gruppe/SeksjonGruppe';
 import Utdanning from './Utdanning';
 import { Element, Undertittel } from 'nav-frontend-typografi';
 import { hentTekst } from '../../../../utils/søknad';
-import { ISpørsmål } from '../../../../models/spørsmalogsvar';
+import { ISpørsmål, ISvar } from '../../../../models/spørsmalogsvar';
 import { utdanningEtterGrunnskolenSpm } from './UtdanningConfig';
 import { useIntl } from 'react-intl';
 import { lagTomUtdanning } from '../../../../helpers/utdanning';
+import { hentBooleanFraValgtSvar } from '../../../../utils/spørsmålogsvar';
 
 interface Props {
   underUtdanning: IUnderUtdanning;
@@ -50,8 +51,10 @@ const TidligereUtdanning: React.FC<Props> = ({
 
   const settHarTattUtdanningEtterGrunnskolen = (
     spørsmål: ISpørsmål,
-    svar: boolean
+    valgtSvar: ISvar
   ) => {
+    const svar: boolean = hentBooleanFraValgtSvar(valgtSvar);
+
     const tattUtdanningEtterGrunnskolenFelt = {
       label: hentTekst(spørsmål.tekstid, intl),
       verdi: svar,
